+++ mjml/helpers/__init__.py
@@ -1,6 +1,7 @@
 
 from .conditional_tag import *
 from .convert_booleans_on_attrs import *
+from .css_inlining import *
 from .fonts import *
 from .format_attributes import *
 from .json_to_xml import *